Adding categories, also manufacturer API call
need to add the links, but the API call involves category_id now in the
filters, and displays the associated category with the product. I also
added in the Controller and Service a way to get a list of
manufacturers. However, I don't believe this will be necessary and I'll
likely remove it.